requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
et-xmlfile==1.1.0
fonttools==4.44.1
idna==3.4
joblib==1.3.2
kiwisolver==1.4.5
lxml==4.9.3
matplotlib==3.8.1
nltk==3.8.1
numpy==1.26.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-pptx==0.6.23
pytz==2023.3.post1
regex==2023.10.3
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
six==1.16.0
sklearn-features==0.0.2
threadpoolctl==3.2.0
tqdm==4.66.1
tzdata==2023.3
urllib3==2.1.0
vaderSentiment==3.3.2
XlsxWriter==3.1.9
